+++ wled00/colors.cpp
@@ -17,7 +17,7 @@ void relativeChangeWhite(int8_t amount, byte lowerBoundary)
   if (new_val > 0xFF) new_val = 0xFF;
   else if (new_val < lowerBoundary) new_val = lowerBoundary;
   col[3] = new_val;
-  colorChanged = true;
+  stateChanged = true;
 }
 
 void colorHStoRGB(uint16_t hue, byte sat, byte* rgb) //hue, sat to rgb