+++ wled00/set.cpp
@@ -550,14 +550,6 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
   byte selectedSeg = strip.getMainSegmentId();
   if (selectedSeg != prevMain) setValuesFromMainSeg();
 
-  //temporary values, do not write direcly to global values if only setting a single segment
-  byte colIn[4]      = {col[0], col[1], col[2], col[3]};
-  byte colInSec[4]   = {colSec[0], colSec[1], colSec[2], colSec[3]};
-  byte effectIn      = effectCurrent;
-  byte speedIn       = effectSpeed;
-  byte intensityIn   = effectIntensity;
-  byte paletteIn     = effectPalette;
-
   bool singleSegment = false;
 
   pos = req.indexOf(F("SS="));
@@ -576,6 +568,16 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
     if (t == 2) for (uint8_t i = 0; i < strip.getMaxSegments(); i++) strip.getSegment(i).setOption(SEG_OPTION_SELECTED, 0); // unselect other segments
     selseg.setOption(SEG_OPTION_SELECTED, t);
   }
+
+  // temporary values, write directly to segments, globals are updated by setValuesFromMainSeg()
+  uint32_t col0    = selseg.colors[0];
+  uint32_t col1    = selseg.colors[1];
+  byte colIn[4]    = {R(col0), G(col0), B(col0), W(col0)};
+  byte colInSec[4] = {R(col1), G(col1), B(col1), W(col1)};
+  byte effectIn    = selseg.mode;
+  byte speedIn     = selseg.speed;
+  byte intensityIn = selseg.intensity;
+  byte paletteIn   = selseg.palette;
 
   uint16_t startI = selseg.start;
   uint16_t stopI  = selseg.stop;
@@ -642,20 +644,17 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
   //set brightness
   updateVal(&req, "&A=", &bri);
 
-  bool col0Changed = false, col1Changed = false, col2Changed = false;
+  bool col0Changed = false, col1Changed = false;
   //set colors
-  updateVal(&req, "&R=", &colIn[0]);
-  updateVal(&req, "&G=", &colIn[1]);
-  updateVal(&req, "&B=", &colIn[2]);
-  updateVal(&req, "&W=", &colIn[3]);
-  for (byte i=0; i<4; i++) if (colIn[i]!=col[i]) col0Changed = colorChanged = true;
-  if (col0Changed) selseg.setColor(0, RGBW32(colIn[0], colIn[1], colIn[2], colIn[3]), selectedSeg); // use transitions
-  updateVal(&req, "R2=", &colInSec[0]);
-  updateVal(&req, "G2=", &colInSec[1]);
-  updateVal(&req, "B2=", &colInSec[2]);
-  updateVal(&req, "W2=", &colInSec[3]);
-  for (byte i=0; i<4; i++) if (colInSec[i]!=colSec[i]) col1Changed = colorChanged = true;
-  if (col1Changed) selseg.setColor(1, RGBW32(colInSec[0], colInSec[1], colInSec[2], colInSec[3]), selectedSeg); // use transitions
+  col0Changed |= updateVal(&req, "&R=", &colIn[0]);
+  col0Changed |= updateVal(&req, "&G=", &colIn[1]);
+  col0Changed |= updateVal(&req, "&B=", &colIn[2]);
+  col0Changed |= updateVal(&req, "&W=", &colIn[3]);
+
+  col1Changed |= updateVal(&req, "R2=", &colInSec[0]);
+  col1Changed |= updateVal(&req, "G2=", &colInSec[1]);
+  col1Changed |= updateVal(&req, "B2=", &colInSec[2]);
+  col1Changed |= updateVal(&req, "W2=", &colInSec[3]);
 
   #ifdef WLED_ENABLE_LOXONE
   //lox parser
@@ -665,6 +664,7 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
     if (parseLx(lxValue, colIn)) {
       bri = 255;
       nightlightActive = false; //always disable nightlight when toggling
+      col0Changed = true;
     }
   }
   pos = req.indexOf(F("LY=")); // Lox secondary color
@@ -673,6 +673,7 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
     if(parseLx(lxValue, colInSec)) {
       bri = 255;
       nightlightActive = false; //always disable nightlight when toggling
+      col1Changed = true;
     }
   }
   #endif
@@ -688,9 +689,7 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
     }
     byte sec = req.indexOf(F("H2"));
     colorHStoRGB(temphue, tempsat, (sec>0) ? colInSec : colIn);
-    if (sec>0) col1Changed = true;
-    else       col0Changed = true;
-    colorChanged = true;
+    col0Changed |= (!sec); col1Changed |= sec;
   }
 
   //set white spectrum (kelvin)
@@ -698,9 +697,7 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
   if (pos > 0) {
     byte sec = req.indexOf(F("K2"));
     colorKtoRGB(getNumVal(&req, pos), (sec>0) ? colInSec : colIn);
-    if (sec>0) col1Changed = true;
-    else       col0Changed = true;
-    colorChanged = true;
+    col0Changed |= (!sec); col1Changed |= sec;
   }
 
   //set color from HEX or 32bit DEC
@@ -708,20 +705,20 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
   pos = req.indexOf(F("CL="));
   if (pos > 0) {
     colorFromDecOrHexString(colIn, (char*)req.substring(pos + 3).c_str());
-    selseg.setColor(0, RGBW32(colIn[0], colIn[1], colIn[2], colIn[3]), selectedSeg); // defined above (SS= or main)
-    col0Changed = colorChanged = true;
+    col0Changed = true;
   }
   pos = req.indexOf(F("C2="));
   if (pos > 0) {
     colorFromDecOrHexString(colInSec, (char*)req.substring(pos + 3).c_str());
-    selseg.setColor(1, RGBW32(colInSec[0], colInSec[1], colInSec[2], colInSec[3]), selectedSeg); // defined above (SS= or main)
-    col1Changed = colorChanged = true;
+    col1Changed = true;
   }
   pos = req.indexOf(F("C3="));
   if (pos > 0) {
     colorFromDecOrHexString(tmpCol, (char*)req.substring(pos + 3).c_str());
-    selseg.setColor(2, RGBW32(tmpCol[0], tmpCol[1], tmpCol[2], tmpCol[3]), selectedSeg); // defined above (SS= or main)
-    col2Changed = colorChanged = true;
+    uint32_t col2 = RGBW32(tmpCol[0], tmpCol[1], tmpCol[2], tmpCol[3]);
+    selseg.setColor(2, col2, selectedSeg); // defined above (SS= or main)
+    stateChanged = true;
+    if (!singleSegment) strip.setColor(2, col2); // will set color to all active & selected segments
   }
 
   //set to random hue SR=0->1st SR=1->2nd
@@ -729,9 +726,7 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
   if (pos > 0) {
     byte sec = getNumVal(&req, pos);
     setRandomColor(sec? colInSec : colIn);
-    if (sec>0) col1Changed = true;
-    else       col0Changed = true;
-    colorChanged = true;
+    col0Changed |= (!sec); col1Changed |= sec;
   }
 
   //swap 2nd & 1st
@@ -743,33 +738,44 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
       colIn[i]    = colInSec[i];
       colInSec[i] = temp;
     }
-    col0Changed = col1Changed = colorChanged = true;
+    col0Changed = col1Changed = true;
   }
 
-  //set effect parameters
-  if (updateVal(&req, "FX=", &effectIn, 0, strip.getModeCount()-1) && request != nullptr) unloadPlaylist();  //unload playlist if changing FX using web request
-  updateVal(&req, "SX=", &speedIn);
-  updateVal(&req, "IX=", &intensityIn);
-  updateVal(&req, "FP=", &paletteIn, 0, strip.getPaletteCount()-1);
-  strip.setMode(selectedSeg, effectIn);
-  selseg.speed     = speedIn;
-  selseg.intensity = intensityIn;
-  selseg.palette   = paletteIn;
-  if (effectIn != effectCurrent || speedIn != effectSpeed || intensityIn != effectIntensity || paletteIn != effectPalette) effectChanged = true;
+  // apply colors to selected segment, and all selected segments if applicable
+  if (col0Changed) {
+    stateChanged = true;
+    uint32_t colIn0 = RGBW32(colIn[0], colIn[1], colIn[2], colIn[3]);
+    selseg.setColor(0, colIn0, selectedSeg);
+    if (!singleSegment) strip.setColor(0, colIn0); // will set color to all active & selected segments
+  }
 
-  //apply to all selected manually to prevent #1618.
-  if (!singleSegment) {
-    for (uint8_t i = 0; i < strip.getMaxSegments(); i++) {
-      WS2812FX::Segment& seg = strip.getSegment(i);
-      if (!seg.isActive() || !seg.isSelected() || i == selectedSeg) continue;
-      if (effectCurrent   != effectIn)    strip.setMode(i, effectIn);
-      if (effectSpeed     != speedIn)     seg.speed     = speedIn;
-      if (effectIntensity != intensityIn) seg.intensity = intensityIn;
-      if (effectPalette   != paletteIn)   seg.palette   = paletteIn;
-      if (col0Changed) seg.setColor(0, RGBW32(colIn[0],       colIn[1],    colIn[2],    colIn[3]), i); // use transitions
-      if (col1Changed) seg.setColor(1, RGBW32(colInSec[0], colInSec[1], colInSec[2], colInSec[3]), i); // use transitions
-      if (col2Changed) seg.colors[2] = RGBW32(tmpCol[0],     tmpCol[1],   tmpCol[2],   tmpCol[3]);
-    }
+  if (col1Changed) {
+    stateChanged = true;
+    uint32_t colIn1 = RGBW32(colInSec[0], colInSec[1], colInSec[2], colInSec[3]);
+    selseg.setColor(1, colIn1, selectedSeg);
+    if (!singleSegment) strip.setColor(1, colIn1); // will set color to all active & selected segments
+  }
+
+  bool fxModeChanged = false, speedChanged = false, intensityChanged = false, paletteChanged = false;
+  // set effect parameters
+  if (updateVal(&req, "FX=", &effectIn, 0, strip.getModeCount()-1)) {
+    if (request != nullptr) unloadPlaylist(); // unload playlist if changing FX using web request
+    fxModeChanged = true;
+  }
+  speedChanged     = updateVal(&req, "SX=", &speedIn);
+  intensityChanged = updateVal(&req, "IX=", &intensityIn);
+  paletteChanged   = updateVal(&req, "FP=", &paletteIn, 0, strip.getPaletteCount()-1);
+  
+  stateChanged |= (fxModeChanged || speedChanged || intensityChanged || paletteChanged);
+
+  //apply to main and all selected segments to prevent #1618.
+  for (uint8_t i = 0; i < strip.getMaxSegments(); i++) {
+    WS2812FX::Segment& seg = strip.getSegment(i);
+    if (i != selectedSeg && (singleSegment || !seg.isActive() || !seg.isSelected())) continue; // skip non main segments if not applying to all
+    if (fxModeChanged)    strip.setMode(i, effectIn);
+    if (speedChanged)     seg.speed     = speedIn;
+    if (intensityChanged) seg.intensity = intensityIn;
+    if (paletteChanged)   seg.palette   = paletteIn;
   }
   setValuesFromMainSeg(); // will fill col[] and cloSec[] as well as effectCurrent, ...
 
@@ -899,16 +905,16 @@ bool handleSet(AsyncWebServerRequest *request, const String& req, bool apply)
   if (pos > 0) {
     userVar1 = getNumVal(&req, pos);
   }
-  //you can add more if you need
+  // you can add more if you need
 
-  if (!apply) return true; //when called by JSON API, do not call colorUpdated() here
+  if (!apply) return true; // when called by JSON API, do not call stateUpdated() here
   
   //internal call, does not send XML response
   pos = req.indexOf(F("IN"));
   if (pos < 1) XML_response(request);
 
   pos = req.indexOf(F("&NN")); //do not send UDP notifications this time
-  colorUpdated((pos > 0) ? CALL_MODE_NO_NOTIFY : CALL_MODE_DIRECT_CHANGE);
+  stateUpdated((pos > 0) ? CALL_MODE_NO_NOTIFY : CALL_MODE_DIRECT_CHANGE);
 
   return true;
 }